flytt logikk over i tagline-komponenten
