+++ src/components/_common/alternativeAudience/AlternativeAudience.tsx
@@ -3,8 +3,6 @@ import { BodyLong } from '@navikt/ds-react';
 import {
     AlternativeAudience as AlternativeAudienceType,
     Audience,
-    AudienceOptions,
-    getAudience,
 } from 'types/component-props/_mixins';
 import { usePageContentProps } from 'store/pageContext';
 import { Language, translator } from 'translations';
@@ -71,45 +69,16 @@ export const AlternativeAudience = () => {
     const { showProductName } = config;
     const { alternativeAudience } = data;
 
-    const getAudienceLabel = translator('audience', language);
-    const getProviderAudienceLabel = translator('providerAudience', language);
     const getStringPart = translator('stringParts', language);
     const getRelatedString = translator('related', language);
-
-    const getProviderTypes = (audience: AudienceOptions) => {
-        if (audience._selected !== Audience.PROVIDER) {
-            return [];
-        }
-        return audience[audience._selected].provider_audience;
-    };
-
-    const buildAudienceAffirmation = (addPeriod: boolean) => {
-        const { audience: currentAudience } = data;
-        const currentAudienceKey = getAudience(currentAudience);
-
-        if (!currentAudience || !currentAudienceKey || currentAudienceKey === 'person') {
-            return '';
-        }
-
-        const currentAudienceLabel = getAudienceLabel(currentAudienceKey);
-        const providerTypes = getProviderTypes(currentAudience) || [];
-        const providerTypesString = joinWithConjunction(
-            providerTypes.map((type) => getProviderAudienceLabel(type)),
-            language
-        );
-
-        const forString = `${getStringPart('for').charAt(0).toUpperCase()}${getStringPart(
-            'for'
-        ).slice(1)}`;
-
-        return `${forString} ${providerTypesString || currentAudienceLabel}${addPeriod ? '.' : ''} `;
-    };
 
     if (!alternativeAudience) {
         return (
             <div className={style.alternativeAudience}>
                 <BodyLong size="small" className={style.text}>
-                    {buildAudienceAffirmation(false)}
+                    {/* {buildAudienceAffirmation(false)}
+                     TODO return ingenting?*/}
+                    Test
                 </BodyLong>
             </div>
         );
@@ -122,7 +91,6 @@ export const AlternativeAudience = () => {
     return (
         <div className={style.alternativeAudience}>
             <BodyLong size="small" className={style.text}>
-                {buildAudienceAffirmation(true)}
                 {getRelatedString('relatedAudience').replace('{name}', productName)}{' '}
                 {audienceLinks.map((link, index) => (
                     <Fragment key={index}>
